Gallery page content

Create the content of a gallery page, comprising three <div> elements: one initial wrapping around before, one main comprising the arbitrary htmltools::tagList(...), and one final wrapping around after. The three elements have custom classes "gallery-before", "gallery-main", "gallery-after", and are wrapped in a parent <div> with class "gallery-container" class plus the provided custom class.

gallery_content(..., before = NULL, after = NULL, class = NULL)

Arguments

...
Unnamed items included in the main <div>

before, after
The content of the <div> before and after the main. If character, the value is wrapped inside htmltools::HTML().

class
Character vector of custom classes.

Value

An HTML tag object that can be rendered as HTML using as.character().